Call julia functions.

julia_call calls julia functions.

julia_call(func_name, ..., need_return = TRUE)

Arguments

func_name
the name of julia function you want to call.

...
the unnamed arguments you want to pass to the julia function.

need_return
whether you want the julia return value or not.

Examples

if (julia_check()) {
  

#> Error: <text>:4:0: unexpected end of input
#> 2: if (julia_check()) {
#> 3:   
#>   ^
# NOT RUN {
 ## julia_setup is quite time consuming
  julia <- julia_setup()

  julia_call("sqrt", 2)

# }
}

#> Error: <text>:1:1: 意外的'}'
#> 1: }
#>     ^